Added submit button to base component
